data-store/pyproject.toml
[tool.poetry]
name = "data-store"
version = "0.1.0"
description = "FastAPI service for storing document metadata"
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.13"
fastapi = "*"
uvicorn = {extras = ["standard"], version = "*"}
sqlalchemy = "*"
alembic = "*"
psycopg2-binary = "*"
pydantic = "*"
pydantic-settings = "^2.1.0"
opentelemetry-api = "^1.27.0"
opentelemetry-sdk = "^1.27.0"
opentelemetry-exporter-otlp = "^1.27.0"
opentelemetry-instrumentation-fastapi = "^0.48b0"
opentelemetry-instrumentation-sqlalchemy = "^0.48b0"

[tool.poetry.group.dev.dependencies]
pytest = "*"
pytest-asyncio = "*"
httpx = "*"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
